run_make_support: move `handle_failed_output` into `util` module
