+++ src/tools/run-make-support/src/path_helpers.rs
@@ -5,7 +5,7 @@ use std::path::{Path, PathBuf};
 
 use crate::command::Command;
 use crate::env_checked::env_var;
-use crate::handle_failed_output;
+use crate::util::handle_failed_output;
 
 /// Return the current working directory.
 ///
